+++ controllers/profiles.js
@@ -6,22 +6,22 @@ const { hashDecode } = require('../utils/hashIdGenerator');
 
 /**
  * @desc    Get all profiles
- * @route   GET /api/profiles
+ * @route   GET /api/users/profiles
  * @access  Public
  */
-exports.getProfiles = asyncHandler(async (req, res) => {
+exports.getProfiles = asyncHandler(async (req, res, next) => {
   res.status(200).json(res.advancedResults);
 });
 
 /**
  * @desc    Get single profile by profile id
- * @route   GET /api/profiles/:id/raw
+ * @route   GET /api/users/:user_id/profiles/raw
  * @access  Public
  */
 exports.getProfile = asyncHandler(async (req, res) => {
   const rows = await db.one(
     'SELECT * FROM profiles WHERE user_id = $1',
-    req.params.id
+    req.params.user_id
   );
   res.status(200).json({
     success: true,
@@ -31,11 +31,16 @@ exports.getProfile = asyncHandler(async (req, res) => {
 
 /**
  * @desc    Get single profile by hash id
- * @route   GET /api/profiles/:hashId
+ * @route   GET /api/users/:user_id/profiles
  * @access  Public
  */
 exports.getProfileByHashId = asyncHandler(async (req, res, next) => {
-  const decodedId = hashDecode(req.params.hashId)[0];
+  // if not user_id params, go to next middleware
+  if (!req.params.user_id) {
+    return next();
+  }
+
+  const decodedId = hashDecode(req.params.user_id)[0];
   if (!decodedId) {
     return next(new ErrorResponse('Invalid user ID', 400));
   }
@@ -43,7 +48,7 @@ exports.getProfileByHashId = asyncHandler(async (req, res, next) => {
     'SELECT * FROM profiles WHERE user_id = $1',
     decodedId
   );
-  res.status(200).json({
+  return res.status(200).json({
     success: true,
     data: rows,
   });
@@ -51,14 +56,14 @@ exports.getProfileByHashId = asyncHandler(async (req, res, next) => {
 
 /**
  * @desc    Update single profile
- * @route   PUT /api/profiles/:id
+ * @route   PUT /api/users/:user_id/profiles
  * @access  Admin/Private
  */
 exports.updateProfile = asyncHandler(async (req, res, next) => {
   // if non-admin user, throw 403 if not updating self
   if (
     req.user.role !== 'admin' &&
-    req.user.user_id.toString() !== req.params.id
+    req.user.user_id.toString() !== req.params.user_id
   ) {
     return next(
       new ErrorResponse(`Not allowed to update other user's profile`, 403)
@@ -97,7 +102,7 @@ exports.updateProfile = asyncHandler(async (req, res, next) => {
     data,
     'profiles',
     'WHERE user_id = $1 RETURNING $2:name',
-    [req.params.id, data]
+    [req.params.user_id, data]
   );
 
   const rows = await db.one(updateProfileQuery);
@@ -110,14 +115,14 @@ exports.updateProfile = asyncHandler(async (req, res, next) => {
 
 /**
  * @desc    Verify single profile
- * @route   PUT /api/profiles/:id/verify
+ * @route   PUT /api/users/:user_id/profiles/verify
  * @access  Admin
  */
 exports.verifyProfile = asyncHandler(async (req, res, next) => {
   // check if user exists
   const user = await db.one(
     'SELECT * FROM profiles WHERE user_id = $1',
-    req.params.id
+    req.params.user_id
   );
 
   const { is_verified } = req.body;
@@ -132,7 +137,7 @@ exports.verifyProfile = asyncHandler(async (req, res, next) => {
     data,
     'profiles',
     'WHERE user_id = $1 RETURNING $2:name',
-    [req.params.id, data]
+    [req.params.user_id, data]
   );
 
   const rows = await db.one(updateIsVerifiedQuery);
@@ -145,14 +150,14 @@ exports.verifyProfile = asyncHandler(async (req, res, next) => {
 
 /**
  * @desc    Upload new or update profile picture
- * @route   PUT /api/profiles/:id/photo
+ * @route   PUT /api/users/:user_id/profiles/upload-photo
  * @access  Admin/Private
  */
 exports.uploadPic = asyncHandler(async (req, res, next) => {
   // if non-admin user, throw 403 if not updating self
   if (
     req.user.role !== 'admin' &&
-    req.user.user_id.toString() !== req.params.id
+    req.user.user_id.toString() !== req.params.user_id
   ) {
     return next(
       new ErrorResponse(
@@ -172,7 +177,7 @@ exports.uploadPic = asyncHandler(async (req, res, next) => {
     data,
     'profiles',
     'WHERE user_id = $1 RETURNING profile_picture',
-    req.params.id
+    req.params.user_id
   );
 
   const rows = await db.one(updateProfileQuery);